backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.1.1
pydantic==2.11.7
pandas==1.5.3
scikit-learn==1.7.1
numpy==1.24.4
python-multipart==0.0.20